modified the fetchNft method in alchemy-sdk.js, called the alchemy fetchNft method in the LLoan and Stake modal, modified the fectchNft method in the MyNfts modal
