+++ src/common/Modals/NftLoansModals/NftLoansModal.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, {useEffect, useState} from "react";
 import "./NftLoansModal.css";
 import "../NftModals/NftModals.css";
 import CancelOutlinedIcon from "@mui/icons-material/CancelOutlined";
@@ -8,10 +8,11 @@ import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import { ethers } from "ethers";
 import contractABI from "../../../utils/ZinarNFTtest.json"
+import fetchNFTsForContract from "../../../services/alchemy-sdk";
 
 const NftLoansModal = (props) => {
   const CONTRACT_ADDRESS = "0x161ed8dc509bdae1b7faaad5b48269bc7c283c05";
-
+  const [zinarNft, setZinarNft] = useState([]);
   const { setNftLoansModal } = props;
 
   const settings = {
@@ -23,6 +24,15 @@ const NftLoansModal = (props) => {
     slidesToScroll: 1,
     initialSlide: 0,
   };
+
+  const getNfts = async () => {
+    const nftData = await fetchNFTsForContract();
+    setZinarNft(nftData);
+  };
+
+  useEffect(() => {
+    getNfts();
+  }, []);
 
   const zinarLoan = async() => {
     const { ethereum } = window;
@@ -64,14 +74,16 @@ const NftLoansModal = (props) => {
         </div>
         <div className="modalContent">
           <Slider {...settings}>
-            {nfts.map((nft) => (
+            {zinarNft.map((nft) => (
               <div className="nftLoansModal">
                 <div className="nftLoansModalTitle">{nft.name}</div>
                 <div className="nftLoansModalContent">
-                  <div>{nft.image}</div>
+                <div key={nft.id}>
+                  <video autoPlay loop src={nft.image} width={250} height={250}/>
+                </div>
                   <div>
                     <div>BUSD Balance</div>
-                    <div>Eligible NFT Balance</div>
+                    <div>NFT ID: {nft.id}</div>
                     <div>Loaned Amount</div>
                     <div>Accrued Interest</div>
                     <button>Borrow Now</button>
